fix: Harden input validation across dictionary, lattice, UTF-8, and pretokenizer
Strengthen defensive checks throughout the codebase to reject malformed or
out-of-range inputs before they can cause undefined behaviour or silent
data corruption:

Lattice (src/core/lattice.cpp):
- Add isValidPos / isValidExtendedPos helpers (anonymous namespace)
- addEdge now rejects zero-length spans (start >= end) and out-of-range
  POS / ExtendedPOS values instead of silently inserting bad nodes

BinaryDictionary (src/dictionary/binary_dict.cpp, binary_dict.h):
- parseData rejects empty surface, out-of-range POS, and (for v0.1+)
  out-of-range ExtendedPOS in each entry record
- BinaryDictWriter::build rejects empty surface and invalid POS /
  ExtendedPOS before writing any bytes
- Fix trailing-space alignment in BinaryDictEntry field comments

DoubleArray (src/dictionary/double_array.cpp):
- build(int32_t) rejects negative values
- build(uint32_t) rejects values that exceed INT32_MAX to avoid
  signed-overflow on cast
- deserialize: add nullptr guard; replace the overflow-prone
  expected-size multiplication with a division-based bounds check;
  load into a temporary vector before overwriting units_ so an
  oversized num_units claim cannot corrupt existing state

DictStringPool (src/dictionary/string_pool.cpp, string_pool.h):
- Widen lengths_ element type from uint16_t to uint32_t so strings
  longer than 65535 bytes are not silently truncated
- add() returns UINT32_MAX sentinel when the pool would exceed uint32_t
  capacity limits
- serialize() returns empty vector when counts exceed uint32_t range
- loadFromMemory: add nullptr guard; use division-based overflow-safe
  size checks; validate every (offset, length) pair against data_size
  before committing; load into temporaries and move-assign on success
  to preserve existing pool state on failure

UserDictionary (src/dictionary/user_dict.cpp, user_dict.h):
- addEntry silently drops entries with empty surface or invalid POS /
  ExtendedPOS values

UTF-8 codec (src/normalize/utf8.cpp):
- decodeUtf8: reject overlong 2-byte sequences (byte1 < 0xC2), 3-byte
  overlong / surrogate-half sequences (E0 + byte2 < A0, ED + byte2 >= A0),
  and 4-byte overlong / above-Unicode-max sequences, returning U+FFFD
- encodeUtf8: guard against surrogate code points (U+D800-U+DFFF) and
  values above U+10FFFF by substituting U+FFFD
- isValidUtf8: rewrite with per-byte dispatch to correctly reject overlong
  sequences, surrogate halves, and code points above U+10FFFF; remove the
  stale comment that called the implementation lenient

Pretokenizer (src/pretokenizer/pretokenizer.cpp):
- tryMatchDate: replace parseDigits with parseInteger so a decimal point
  (e.g. "20.24年") cannot satisfy the year/month/day digit parser

Documentation (src/dictionary/core_dict.h, dictionary.h, user_dict.h):
- Clarify start_pos parameter as byte offset rather than character index

Testing:
- binary_dict_test: 6 new cases for empty surface, invalid POS / ExtendedPOS
  on both load and build paths
- dict_string_pool_test: 5 new cases covering nullptr input, huge count,
  out-of-range string reference, long-string round-trip (>65535 bytes)
- double_array_test: 4 new cases for negative values, uint32 overflow,
  UINT32_MAX num_units without buffer overflow, and failure-preserves-state
- user_dict_test: 1 new case verifying invalid entries are silently rejected
- utf8_test: replace lenient-behaviour stubs with strict assertions for
  overlong 2/3/4-byte, surrogate-half, above-U+10FFFF, and encode-guard cases
- pretokenizer_number_test: 1 new case ensuring a decimal-point year is not
  recognised as a Date token

+++ src/dictionary/string_pool.h
@@ -70,7 +70,7 @@ class DictStringPool {
  private:
   std::string data_;               // Concatenated strings
   std::vector<uint32_t> offsets_;  // Start offset for each string
-  std::vector<uint16_t> lengths_;  // Length of each string
+  std::vector<uint32_t> lengths_;  // Length of each string
 };
 
 }  // namespace suzume::dictionary
